Debug logs for /ask (#43)
Added a new param to the "ask" call. This param will trigger logs to be emitted by the backend.

We'll use this as an early method to be able to measure our prompts' effectiveness and track improvements over time.

Oh, you don't like the GET param? Neither do I. But it's a nice way to introduce a parameter without updating specs, which means we can remove it safely once the log feature is done for real.